cloud-functions/us-patent-process/pyproject.toml
[tool.poetry]
name = "us-patent-process"
version = "0.1.0"
description = ""
authors = ["Marcus Vinicius Monteiro de Souza <marcusmonteirodesouza@gmail.com>"]
readme = "README.md"
packages = [{include = "us_patent_process"}]

[tool.poetry.dependencies]
python = "^3.10"
cloudevents = "^1.9.0"
functions-framework = "^3.4.0"
google-cloud-documentai = "^2.18.0"
google-cloud-storage = "^2.10.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
